src/task_generator.py
```python
from datetime import datetime
from src.database_manager import DATABASE

class Task:
    def __init__(self, id, description, priority, status, created_at):
        self.id = id
        self.description = description
        self.priority = priority
        self.status = status
        self.created_at = created_at

def generateTask(user_input, bot_output):
    """
    Function to generate tasks based on user inputs and bot outputs.
    """
    # Analyze the bot output to identify tasks
    tasks = analyzeBotOutput(bot_output)

    # Prioritize tasks based on user input
    prioritized_tasks = prioritizeTasks(tasks, user_input)

    # Save tasks to the database
    saveTasks(prioritized_tasks)

def analyzeBotOutput(bot_output):
    """
    Function to analyze bot output and identify tasks.
    """
    # Placeholder for task identification logic
    tasks = []

    # Return identified tasks
    return tasks

def prioritizeTasks(tasks, user_input):
    """
    Function to prioritize tasks based on user input.
    """
    # Placeholder for task prioritization logic
    prioritized_tasks = []

    # Return prioritized tasks
    return prioritized_tasks

def saveTasks(tasks):
    """
    Function to save tasks to the database.
    """
    for task in tasks:
        DATABASE.insert('tasks', {
            'id': task.id,
            'description': task.description,
            'priority': task.priority,
            'status': task.status,
            'created_at': task.created_at
        })
```